% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust2.R
\name{hclust2}
\alias{hclust2}
\title{Faster Hierarchical Clustering in Pseudometric Spaces}
\usage{
hclust2(d = NULL, method = c("single", "gini", "exemplar", "exemplar2",
  "exemplar_approx", "exemplar_naive"), objects = NULL, ...)
}
\arguments{
\item{d}{an object of class \code{\link{dist}},
\code{NULL}, an R function, or a single string}

\item{method}{a single string,
one of: \code{single}, \code{gini}}

\item{objects}{\code{NULL}, numeric matrix, a list, or a character vector}

\item{...}{internal tuning parameters,
e.g., \code{control=list(thresholdGini=0.3)} for the \code{gini} method.}
}
\value{
A named list of class \code{hclust}, see \code{\link{hclust}},
with additional components:
\itemize{
     \item stats   performance statistics
     \item control internal tuning parameters used
}
}
\details{
For compatibility with \code{\link{hclust}}, \code{d} may be an object
of class \code{dist}. In such a case, the \code{objects} argument is ignored.

If \code{d} is an R function, then \code{objects} should be an R list.
Here, \code{d(objects[[i]], objects[[j]])} gives the value of a pseudometric.

If \code{objects} is a character vector or a list, then \code{d}
should be a single string, one of: \code{levenshtein} (or \code{NULL}),
\code{dinu}, \code{hamming}, \code{euclinf}. Note that the list must consist either
of integer or of numeric vectors only (depending on the pseudometric of choice).
Each string must be in ASCII, but you can always convert it to UTF-32 with
\link[stringi]{stri_enc_toutf32}.

Otherwise, if \code{objects} is a numeric matrix, then \code{d} should be
a single string, one of: \code{euclidean} (or \code{NULL}),
\code{manhattan}, \code{maximum}, or \code{hamming}.
}

