% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/stack.R
\name{stack_create}
\alias{Stack}
\alias{as.list.Stack}
\alias{format.Stack}
\alias{print.Stack}
\alias{stack_create}
\alias{stack_empty}
\alias{stack_pop}
\alias{stack_push}
\alias{str.Stack}
\title{Stack}
\usage{
stack_create()

\method{as.list}{Stack}(stack)

stack_empty(stack)

stack_push(stack, obj)

stack_pop(stack)

\method{print}{Stack}(stack, ...)

\method{format}{Stack}(stack, ...)

\method{str}{Stack}(stack, ...)
}
\arguments{
\item{stack}{a stack object}

\item{obj}{an R object}

\item{...}{other arguments passed to or from other methods}

\item{...}{other arguments passed to or from other methods}
}
\value{
\code{stack_create} returns a new, empty stack.

\code{as.list.Stack} returns an R list object.

\code{stack_empty} returns a single logical value.

\code{stack_push} does not return anything interesting.

\code{stack_pop} returns an object at the top of te stack or
throws an error if the stack is empty.

\code{print.Stack} returns the stack coerced to a list, invisibly.

\code{format.Stack} returns the character representation
of the objects in the stack.

\code{str.Stack} does not return anything interesting.
}
\description{
Stack is an abstract data type which allows for pushing elements
and popping them in reverse (last in-first out) order.
}
\details{
\code{stack_create} creates a new stack.

\code{as.list.Stack} converts a given stack object to an R list.

\code{stack_empty} determines if a given stack is empty or not.

\code{stack_push} pushes a given object to the top of the stack.

\code{stack_pop} pops an object from the top of the stack.

\code{print.Stack} prints the contents of the stack
(coerced to a list object) on the console.

\code{format.Stack} pretty-prints the contents of the stack
(coerced to a list object) on the console.

\code{str.Stack} compactly displays the contents of the stack
(coerced to a list object) on the console.
}

