% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie-package.R
\docType{package}
\name{genie-package}
\alias{genie-package}
\alias{genie}
\title{The Genie Package}
\description{
See \code{\link{hclust2}()} for details.
}
\author{
Marek Gagolewski, Maciej Bartoszuk, Anna Cena
}
