% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hclust2.R
\name{hclust2}
\alias{hclust2}
\title{Faster Hierarchical Clustering in Pseudometric Spaces}
\usage{
hclust2(d = NULL, method = c("single", "complete", "exemplar",
  "exemplar_naive"), objects = NULL, ...)
}
\arguments{
\item{d}{an object of class \code{\link{dist}},
\code{NULL}, an R function or a single string}

\item{method}{a single string,
one of \code{single}, \code{complete}, or \code{exemplar}}

\item{objects}{\code{NULL}, numeric matrix or a list}

\item{...}{internal tuning parameters}
}
\value{
A named list of class \code{hclust}, see \code{\link{hclust}},
with additional components:
\itemize{
\item stats ...
\item control ...
}
}
\description{
Faster Hierarchical Clustering in Pseudometric Spaces
}
\details{
For compatibility with \code{\link{hclust}}, \code{d}
may be an object of class \code{dist}. In such a case, the \code{objects}
argument is ignored.

If \code{d} is an R function, then \code{objects} should be an R list.
Here, \code{d(objects[[i]], objects[[j]])} gives the value of a pseudometric.

If \code{objects} is a character vector, then \code{d} is a single string,
one of: \code{levenshtein} (or \code{NULL}), \code{dinu}.

Otherwise, if \code{objects} is a numeric matrix,
then \code{d} is a single string, one of: \code{euclidean} (or \code{NULL}),
\code{manhattan}, \code{maximum}, or \code{hamming}.
}

